requirements.txt
amqp==1.4.6
ansible==1.9.1
anyjson==0.3.3
astroid==1.3.7
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.19
bulbs==0.3.14.post20130504
celery==3.1.18
certifi==2015.4.28
CherryPy==3.7.0
Cython==0.22
dateutils==0.6.6
Django==1.9.1
django-braces==1.4.0
django-celery==3.0.17
django-cors-headers==1.1.0
django-debug-toolbar==0.9.4
django-extensions==1.1.1
django-filter==0.12.0
django-floppyforms==1.1
django-haystack==2.0.0
django-mediagenerator==1.11
django-oauth-toolkit==0.7.2
django-pdb==0.3.2
django-pretty-times==0.1.0
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
docker-compose==1.1.0
docker-py==0.7.2
docker-ship==0.1
dockerpty==0.3.2
docopt==0.6.2
ecdsa==0.13
futures==3.0.5
gcalcli==3.3
GDAL==1.11.3
google-api-python-client==1.2
greenlet==0.4.9
gunicorn==19.1.1
httplib2==0.8
ipython==2.0.0
Jinja2==2.7.3
jsonschema==2.4.0
kombu==3.0.24
logilab-common==1.0.2
Markdoc==0.6.6
Markdown==2.3.1
MarkupSafe==0.23
msgpack-python==0.4.7
neovim==0.1.0
nltk==2.0.4
numpy==1.9.1
oauthlib==0.7.2
omnijson==0.1.2
paramiko==1.15.2
PIL==1.1.7
Pillow==2.6.1
psycopg2==2.5.5
pycrypto==2.6.1
Pygments==2.0.2
PyJWT==1.4.0
pylint==1.4.4
pylint-django==0.6.1
pylint-plugin-utils==0.2.3
pyserial==2.6
python-dateutil==2.2
python-gflags==2.0
pytz==2014.10
PyYAML==3.11
requests==2.4.3
scikit-learn==0.15.2
scipy==0.14.0
six==1.9.0
South==0.8.1
texttable==0.8.3
trollius==2.1
vboxapi==1.0
virtualenv==1.10.1
WebOb==1.4.1
websocket-client==0.26.0
wheel==0.24.0
Whoosh==2.4.1
